+++ packages/insight-viewer/src/hooks/useInteraction/useHandleDrag/control.ts
